requirements.txt
fastapi
uvicorn
python-dotenv
prometheus_fastapi_instrumentator
sentry-sdk
